backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pymupdf==1.24.5
groq==1.0.0
pydantic==2.7.4
python-dotenv==1.0.1
httpx>=0.27.0
